feat(shared): extract shared TypeScript types into @naur/shared package
Closes #177
