+++ pnpm-workspace.yaml
@@ -1,3 +1,4 @@
 packages:
   - "services/naurffxiv"
+  - "services/shared"
   - "tests/e2e"
