+++ services/naurffxiv/src/app/(protected)/dev/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import type { FCWithChildren } from "@/types/common";
+import type { FCWithChildren } from "@naur/shared/types";
 import type { ReactElement } from "react";
 import { useRouter } from "next/navigation";
