+++ services/naurffxiv/src/app/(protected)/mod-portal/layout.tsx
@@ -1,16 +1,11 @@
 "use client";
 
-import React, { ReactNode } from "react";
+import type { WithChildren } from "@naur/shared/types";
+import type { ReactNode } from "react";
 
 import { ModNavigation } from "@/components/ModPortal/ModNavigation/ModNavigation";
 
-type ModPortalLayoutProps = {
-  children: ReactNode;
-};
-
-export default function ModPortalLayout({
-  children,
-}: ModPortalLayoutProps): ReactNode {
+export default function ModPortalLayout({ children }: WithChildren): ReactNode {
   return (
     <div className="grid grid-cols-1 lg:grid-cols-[25ch_1fr] max-w-screen-2xl mx-auto py-4">
       <div className="scrollbar">
