+++ services/naurffxiv/src/components/Providers/MUITheme.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import type { ChildrenProps } from "@/types/common";
-import type { ReactElement } from "react";
+import type { WithChildren } from "@naur/shared/types";
+import type { ReactNode } from "react";
 import { ThemeProvider } from "@emotion/react";
 import { theme } from "@/config/theme";
 
-export default function MUITheme({ children }: ChildrenProps): ReactElement {
+export default function MUITheme({ children }: WithChildren): ReactNode {
   return <ThemeProvider theme={theme}>{children}</ThemeProvider>;
 }
