+++ services/shared/.pre-commit-config.yaml
@@ -1,0 +1,24 @@
+# Shared package configuration
+orphan: true
+repos:
+  - repo: builtin
+    hooks:
+      - id: trailing-whitespace
+      - id: end-of-file-fixer
+
+  - repo: local
+    hooks:
+      - id: prettier
+        name: prettier
+        entry: npx prettier --write --ignore-unknown --cache
+        language: system
+        files: \.(ts|json|yaml|md)$
+        priority: 0
+
+      - id: tsc
+        name: tsc
+        entry: npx tsc --noEmit
+        language: system
+        types: [ts]
+        pass_filenames: false
+        priority: 2
