+++ services/shared/types/auth/index.ts
@@ -1,0 +1,2 @@
+export * from "./error";
+export * from "./events";
