+++ services/shared/types/common.ts
@@ -1,4 +1,4 @@
-import type { ReactElement, ReactNode } from "react";
+import type { ReactNode } from "react";
 
 // Requires children to be present, unlike React's PropsWithChildren
 export type WithChildren = {
@@ -8,7 +8,4 @@ export type WithChildren = {
 // Functional component type with required children
 export type FCWithChildren<P = unknown> = (
   props: P & WithChildren,
-) => ReactElement;
-
-// Alias for WithChildren – useful for destructured props
-export type ChildrenProps = WithChildren;
+) => ReactNode;
