+++ services/shared/types/discord/index.ts
@@ -1,0 +1,2 @@
+export * from "./oauth";
+export * from "./entities";
